feat: make public Cart view
